Use property to update junction name of m2m fields (#2704)
When creating a m2m field and toggling the "Auto-Generate" checkbox in
the "Relation" tab multiple times, `collection_many` in
`relationInfoM2M` is not set correctly. This results in an error when
creating the field with an auto-generated junction collection. This
fixes the behavior by unifying the update of `relationInfoM2M`
between the input component used when auto-generating and the select
component used otherwise.